database/requirements.txt
flask
Pillow
typing
pymongo